Change location tregex rule
The original rule described in the paper did not work for the sample
documents. The original rule `NP < location markers` is so general that
it produces far too many false positives, if good markers are used.
E.g. "institution" needs to be a markers in order to match locations.

The new rule works fairly well, because it looks for words such as "at",
"in" etc. that cooccur with existing markers.